MCOL-4740 UPDATE involving multi-tables returns wrong "Rows matched".
For UPDATEs involving a single table, the server call to
handler::direct_update_rows() is used to correctly set the count for
the number of updated rows in the UPDATE statement.

However, for UPDATEs involving multi-tables, the server does not call
handler::direct_update_rows(). This patch adds support to correctly
report the number of updated rows to the client by setting
multi_update::updated and multi_update::found in handler::rnd_end().